+++ apps/platform/CHANGELOG.md
@@ -1,5 +1,124 @@
 # platform
 
+## 0.0.0-dev-20260505081231
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505081231
+
+## 0.0.0-dev-20260505080921
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505080921
+
+## 0.0.0-dev-20260505080217
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505080217
+
+## 0.0.0-dev-20260505075652
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505075652
+
+## 0.0.0-dev-20260505075404
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505075404
+
+## 0.0.0-dev-20260505075017
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505075017
+
+## 0.0.0-dev-20260505074401
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505074401
+
+## 0.0.0-dev-20260505074040
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505074040
+
+## 0.0.0-dev-20260505073736
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505073736
+
+## 0.0.0-dev-20260505073225
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505073225
+
+## 0.0.0-dev-20260505072907
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505072907
+
+## 0.0.0-dev-20260505072427
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505072427
+
+## 0.0.0-dev-20260505072314
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505072314
+
+## 0.0.0-dev-20260505071855
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505071855
+
+## 0.0.0-dev-20260505070643
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505070643
+
+## 0.0.0-dev-20260505065905
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505065905
+
+## 0.0.0-dev-20260505065504
+
+### Patch Changes
+
+- Updated dependencies
+  - @kalphq/compiler@0.0.0-dev-20260505065504
+
 ## 0.0.0-dev-20260505064336
 
 ### Patch Changes