+++ packages/cloudflare/CHANGELOG.md
@@ -1,5 +1,158 @@
 # @kalphq/cloudflare
 
+## 0.0.0-dev-20260505081231
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505081231
+  - @kalphq/core@0.0.0-dev-20260505081231
+
+## 0.0.0-dev-20260505080921
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505080921
+  - @kalphq/core@0.0.0-dev-20260505080921
+
+## 0.0.0-dev-20260505080217
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505080217
+  - @kalphq/core@0.0.0-dev-20260505080217
+
+## 0.0.0-dev-20260505075652
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505075652
+  - @kalphq/core@0.0.0-dev-20260505075652
+
+## 0.0.0-dev-20260505075404
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505075404
+  - @kalphq/core@0.0.0-dev-20260505075404
+
+## 0.0.0-dev-20260505075017
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505075017
+  - @kalphq/core@0.0.0-dev-20260505075017
+
+## 0.0.0-dev-20260505074401
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505074401
+  - @kalphq/core@0.0.0-dev-20260505074401
+
+## 0.0.0-dev-20260505074040
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505074040
+  - @kalphq/core@0.0.0-dev-20260505074040
+
+## 0.0.0-dev-20260505073736
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505073736
+  - @kalphq/core@0.0.0-dev-20260505073736
+
+## 0.0.0-dev-20260505073225
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505073225
+  - @kalphq/core@0.0.0-dev-20260505073225
+
+## 0.0.0-dev-20260505072907
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505072907
+  - @kalphq/core@0.0.0-dev-20260505072907
+
+## 0.0.0-dev-20260505072427
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505072427
+  - @kalphq/core@0.0.0-dev-20260505072427
+
+## 0.0.0-dev-20260505072314
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505072314
+  - @kalphq/core@0.0.0-dev-20260505072314
+
+## 0.0.0-dev-20260505071855
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505071855
+  - @kalphq/core@0.0.0-dev-20260505071855
+
+## 0.0.0-dev-20260505070643
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505070643
+  - @kalphq/core@0.0.0-dev-20260505070643
+
+## 0.0.0-dev-20260505065905
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505065905
+  - @kalphq/core@0.0.0-dev-20260505065905
+
+## 0.0.0-dev-20260505065504
+
+### Patch Changes
+
+- dev release
+- Updated dependencies
+  - @kalphq/sdk@0.0.0-dev-20260505065504
+  - @kalphq/core@0.0.0-dev-20260505065504
+
 ## 0.0.0-dev-20260505064336
 
 ### Patch Changes